+++ src/index.ts
@@ -4,6 +4,8 @@
  */
 
 export { Jexl, default } from './Jexl.ts'
+export type { JexlOptions } from './Jexl.ts'
+export type { GetMember, VariableReader } from './grammar.ts'
 export { default as Expression } from './Expression.ts'
 export { default as Lexer } from './Lexer.ts'
 export { getGrammar } from './grammar.ts'